refactored and continue work on codegen
